+++ src/piping.ts
@@ -92,8 +92,14 @@ const indexPage: string =
   <h1>Piping</h1>
   Streaming file sending/receiving
   <form method="POST" id="file_form" enctype="multipart/form-data">
-    <h3>Step 1: Choose a file</h3>
-    <input type="file" name="input_file"><br>
+    <h3>Step 1: Choose a file or text</h3>
+
+    <input type="checkbox" id="inputMode" onchange="toggleInputMode()">: <b>Text mode</b><br><br>
+
+    <input type="file" name="input_file">
+    <textarea type="text" name="input_text" placeholder="Input text" cols="30" rows="10"></textarea>
+    <br>
+
     <h3>Step 2: Write your secret path</h3>
     (e.g. "abcd1234", "mysecret.png?n=3")<br>
     <input id="secret_path" placeholder="Secret path" size="50"><br>
@@ -106,14 +112,46 @@ const indexPage: string =
     https://github.com/nwtgck/piping-server#readme
   </a><br>
   <script>
-    var fileForm = document.getElementById("file_form");
-    var secretPathInput = document.getElementById("secret_path");
-    secretPathInput.onkeyup = function(){
-      fileForm.action = "/" + secretPathInput.value;
-    };
+    // Set secret path action routing
+    (function () {
+      var fileForm = document.getElementById("file_form");
+      var secretPathInput = document.getElementById("secret_path");
+      secretPathInput.onkeyup = function(){
+        fileForm.action = "/" + secretPathInput.value;
+      };
+    })();
+
+    // Toggle input mode: file or text
+    var toggleInputMode = (function () {
+      var fileInput   = document.getElementsByName("input_file")[0];
+      var textInput   = document.getElementsByName("input_text")[0];
+      var activeInput = fileInput;
+      var deactivatedInput = textInput;
+
+      // Set inputs' functionality and visibility
+      function setInputs() {
+        activeInput.disabled = false;
+        activeInput.style.display = null;
+
+        deactivatedInput.disabled = true;
+        deactivatedInput.style.display = "none";
+      }
+      setInputs();
+
+      // Body of toggleInputMode
+      function toggle() {
+        // Swap inputs
+        var tmpInput     = activeInput;
+        activeInput      = deactivatedInput;
+        deactivatedInput = tmpInput;
+        setInputs();
+      }
+      return toggle;
+    })();
   </script>
 </body>
-</html>`;
+</html>
+`;
 
 /**
  * Generate help page
@@ -304,7 +342,10 @@ export class Server {
               part.byteCount === undefined ?
                 {} : {"Content-Length": part.byteCount}
             ),
-            "Content-Type": part.headers["content-type"]
+            ...(
+              part.headers["content-type"] === undefined ?
+                {} : {"Content-Type": part.headers["content-type"]}
+            )
           };
 
       // Write headers to a receiver